Move the tool helpers next to the tools
Nothing outside src/tools/ calls these -- the commands already hold a
server object where a tool only gets a name from the model. Entry points
come from the manifest's tools array, not from globbing the directory, so
a non-tool file can live here without ray treating it as one.
