+++ src/tools/deploy-site.ts
@@ -1,6 +1,6 @@
 import { Tool } from "@raycast/api";
 import { Site } from "../api/Site";
-import { findSite } from "../lib/resolve";
+import { findSite } from "./helpers";
 
 type Input = {
   /**
